requirements.txt
# Django Core
Django==4.2.25
django-environ==0.12.0

# Database
psycopg2-binary==2.9.11

# Google Drive API
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Excel/Data Processing
pandas==2.2.3
openpyxl==3.1.5
numpy==2.1.3

# Utilities
python-dateutil==2.9.0
tzdata==2025.2

# Development/Testing
Faker==37.11.0

# Web Server (배포용)
gunicorn==23.0.0
whitenoise==6.8.2

# 기타
asgiref==3.10.0
sqlparse==0.5.3
typing_extensions==4.15.0
